Improved DVS code
